requirements.txt
langchain-core>=0.1.0
langchain-deepseek>=0.1.0
python-dotenv>=1.0.0